requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
filelock==3.0.12
idna==2.10
lxml==4.6.3
pdfminer==20191125
Pillow==8.2.0
pycryptodome==3.10.1
PyMuPDF==1.18.14
requests==2.25.1
soupsieve==2.2.1
uritools==3.0.2
urllib3==1.26.6
